Create tests/tests.h
